:lipstick: Split content of interfaces.ts
