+++ src/idata_event.ts
@@ -1,0 +1,5 @@
+import {IEvent} from './ievent';
+
+export interface IDataEvent extends IEvent {
+    data: any;
+}
