+++ src/ierror_event.ts
@@ -1,0 +1,6 @@
+import {IEvent} from './ievent';
+
+export interface IErrorEvent extends IEvent {
+    error: any;
+    message: string;
+}
