+++ src/ievent.ts
@@ -1,0 +1,5 @@
+import {IEventMetadata} from './ievent_metadata';
+
+export interface IEvent {
+    metadata: IEventMetadata;
+}
