+++ src/index.ts
@@ -1,1 +1,7 @@
-export * from './interfaces';
+export * from './idata_event';
+export * from './ientity_event';
+export * from './ierror_event';
+export * from './ievent';
+export * from './ievent_aggregator';
+export * from './ievent_metadata';
+export * from './isubscription';
