--- src/interfaces.ts
@@ -1,39 +1,0 @@
-import {ExecutionContext, IEntity} from '@essential-projects/core_contracts';
-
-export interface IEventAggregator {
-  publish(event: string | any, data?: any): void;
-  subscribe(event: string | Function, callback: Function): ISubscription;
-  subscribeOnce(event: string | Function, callback: Function): ISubscription;
-  createEntityEvent(data: any, source: IEntity, context: ExecutionContext, metadataOptions?: {
-    [key: string]: any,
-  }): IEvent;
-}
-
-export interface ISubscription {
-  dispose(): void;
-}
-
-export interface IEntityEvent extends IDataEvent {
-  source: IEntity;
-}
-
-export interface IDataEvent extends IEvent {
-  data: any;
-}
-
-export interface IErrorEvent extends IEvent {
-  error: any;
-  message: string;
-}
-
-export interface IEvent {
-  metadata: IEventMetadata;
-}
-
-export interface IEventMetadata {
-  id: string;
-  applicationId?: string;
-  context: ExecutionContext;
-  response?: string;
-  options?: { [key: string]: any };
-}
