+++ src/isubscription.ts
@@ -1,0 +1,3 @@
+export interface ISubscription {
+    dispose(): void;
+}
